pyproject.toml
[project]
name = "software-delivery-adw"
version = "0.1.0"
description = "AI Development Workflow for Software Delivery"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "exa-py>=1.0.0",
    "openai>=1.0.0",
    "boto3>=1.40.55",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["adws"]
